Work around build issues when this is used as a dev pod
Unfortunately I am still not quite sure what's the root cause: for some reason
MMMContainerView and MMMLayoutUtils.inverseGolden could not be found when this
pod was included as a development one. Separating them into their own sources
seems to fix the problem for now but I guess it's just a symptom of another
issue.